% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsp_instance.R
\name{as_TSP}
\alias{as_TSP}
\title{Convert to TSP instance object of package TSP.}
\usage{
as_TSP(x)
}
\arguments{
\item{x}{[\code{\link{tsp_instance}}]\cr
TSP instance.}
}
\value{
[\code{\link[TSP]{TSP}}].
}
\description{
Convert to TSP instance object of package TSP.
}

