% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feature_centroid.R
\name{feature_centroid}
\alias{feature_centroid}
\title{Centroid features.}
\usage{
feature_centroid(x)
}
\arguments{
\item{x}{[\code{\link{tsp_instance}}]\cr
TSP instance.}
}
\value{
[\code{list}].
}
\description{
Includes the coordinates of the mean coordinates of the the point cloud
and the statistics of the distances of all cities from it.
}

