\name{feature_modes}
\alias{feature_modes}
\title{Modes of edge cost distribution feature.}
\usage{
feature_modes(x)
}
\arguments{
  \item{x}{[\code{\link{tsp_instance}}]\cr TSP instance.}
}
\value{
[\code{list}] List containing (estimated) number of modes.
}
\description{
Includes the number of modes of the edge cost distribution.
}

