\name{feature_mst}
\alias{feature_mst}
\title{MST features.}
\usage{
feature_mst(x)
}
\arguments{
  \item{x}{[\code{\link{tsp_instance}}]\cr TSP instance.}
}
\value{
[\code{list}].
}
\description{
Construct minimun spanning tree, then calculate the
statistics of a) the distances in the MST, b) the depths of
all nodes in the MST.
}

