% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsp_instance.R
\name{number_of_cities}
\alias{number_of_cities}
\title{Get number of cities in tsp instance.}
\usage{
number_of_cities(x)
}
\arguments{
\item{x}{[\code{\link{tsp_instance}}]\cr
TSP instance.}
}
\value{
[\code{integer(1)}].
}
\description{
Get number of cities in tsp instance.
}

