% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsp_instance.R
\name{print.tsp_instance}
\alias{print.tsp_instance}
\title{Print TSP instance}
\usage{
\method{print}{tsp_instance}(x, ...)
}
\arguments{
\item{x}{[\code{\link{tsp_instance}}]\cr
TSP instance.}

\item{...}{[any]\cr
Not used.}
}
\description{
Print TSP instance
}

