\name{remove_zero_distances}
\alias{remove_zero_distances}
\title{Remove any duplicate cities in a tsp instance.}
\usage{
remove_zero_distances(instance)
}
\arguments{
  \item{instance}{[\code{tsp_instance}]\cr TSP instance
  object.}
}
\value{
New TSP instance in which all duplicate cities have been
removed.
}
\description{
Remove any duplicate cities in a tsp instance.
}

