% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rescale_instance.R
\name{rescale_instance}
\alias{rescale_coords}
\alias{rescale_instance}
\title{Rescale coords of TSP instance to \eqn{[0,1]^2}.}
\usage{
rescale_instance(x)

rescale_coords(coords)
}
\arguments{
\item{x}{[\code{\link{tsp_instance}}]\cr
TSP instance.}

\item{coords}{[\code{matrix}]\cr
Numeric matrix of city coordinates,
rows denote cities.}
}
\value{
[\code{matrix}] for \code{rescale_coords} and \code{tsp_instance} for \code{rescale_instance}.
  Numeric matrix of scaled city coordinates.
}
\description{
Rescale coords of TSP instance to \eqn{[0,1]^2}.
}

