% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsp_instance.R
\name{tsp_instance}
\alias{tsp_instance}
\title{Generates a TSP instance S3 object either from city coordinates.}
\usage{
tsp_instance(coords, dists)
}
\arguments{
\item{coords}{[\code{matrix}]\cr
Numeric matrix of city coordinates,
rows denote cities.}

\item{dists}{[\code{dist}]\cr
Optional distance matrix containing the inter-city distances.
If not provided, the (euclidean) distances are computed from the coordinates.}
}
\value{
[\code{\link{tsp_instance}}].
}
\description{
Generates a TSP instance S3 object either from city coordinates.
}

