\name{feature_chull}
\alias{feature_chull}
\title{Convex hull features.}
\usage{
feature_chull(x)
}
\arguments{
  \item{x}{[\code{\link{tsp_instance}}]\cr TSP instance.}
}
\value{
[\code{list}].
}
\description{
Determines the area of the convex hull and the ratio of the
cities which lie on the convex hull in the euklidean space.
}

