% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feature_distance.R
\name{feature_distance}
\alias{feature_distance}
\title{Distance features.}
\usage{
feature_distance(x)
}
\arguments{
\item{x}{[\code{\link{tsp_instance}}]\cr
TSP instance.}
}
\value{
[\code{list}]
  List of statistics describing the distribution of distances.
}
\description{
Computes different statistics describing the distribution of
pairwise distances between cities.
}

