% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feature_mst.R
\name{feature_mst}
\alias{feature_mst}
\title{MST features.}
\usage{
feature_mst(x)
}
\arguments{
\item{x}{[\code{\link{tsp_instance}}]\cr
TSP instance.}
}
\value{
[\code{list}].
}
\description{
Construct minimun spanning tree, then calculate
the statistics of
a) the distances in the MST,
b) the depths of all nodes in the MST.
}

