% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feature_nnds.R
\name{feature_nnds}
\alias{feature_nnds}
\title{Nearest neighbor features.}
\usage{
feature_nnds(x)
}
\arguments{
\item{x}{[\code{\link{tsp_instance}}]\cr
TSP instance.}
}
\value{
[\code{list}].
}
\description{
Statistics describing the distribution of distances of each city
to its nearest neighbor.
}

