\name{instance_dim}
\alias{instance_dim}
\title{Get instance dimensionality (space where coords live).}
\usage{
instance_dim(x)
}
\arguments{
  \item{x}{[\code{\link{tsp_instance}}]\cr TSP instance.}
}
\value{
[\code{integer(1)}].
}
\description{
Get instance dimensionality (space where coords live).
}

