% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rotation_utilities.R
\name{normalization_angle}
\alias{normalization_angle}
\title{Calculate rotation angle such that the main axis through the
cities is aligned with the X axis.}
\usage{
normalization_angle(instance)
}
\arguments{
\item{instance}{[\code{\link{tsp_instance}}]\cr
  TSP instance.}
}
\value{
[\code{numeric(1)}]
}
\description{
Calculate rotation angle such that the main axis through the
cities is aligned with the X axis.
}

